requirements-dev.txt
-r requirements.txt
coverage==7.6.1
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
ruff==0.7.0
pre-commit==3.8.0
detect-secrets==1.5.0
django-debug-toolbar==4.4.6
